+++ src/Effect/Error.hs
@@ -9,12 +9,14 @@ type MonadEffectError e m = (MonadError e m, From EffectError e)
 
 data EffectError
     = GettingDataDirectoryFailed
+    | GettingConfigDirectoryFailed
     | ReadFailed IOException
     | WriteFailed IOException
     | BackupFailed
 
 instance Show EffectError where
     show GettingDataDirectoryFailed = "Failed to find data storage directory"
+    show GettingConfigDirectoryFailed = "Failed to find configuration storage directory"
     show (ReadFailed e) = "Failed to load data: " <> show (ioeGetErrorType e)
     show (WriteFailed e) = "Failed to save data" <> show (ioeGetErrorType e)
     show BackupFailed = "Failed to backup data"