+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added importance as a task component.
 - Added importance range filtering to list and delete subcommands.
 - Added `list --reverse` option.
+- Added support for user-defined program configuration via `config.toml` (parsed as TOML 0.5).
 
 ### Changed
 - Adds version information to registry serialization. On deserialization failure, the existing data is backed up and a new registry is created.